requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
EasyProcess==1.1
entrypoint2==1.1
frozenlist==1.3.3
idna==3.4
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
numpy==1.24.3
openai==0.27.7
opencv-python==4.7.0.72
packaging==23.1
pefile==2023.2.7
Pillow==9.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytesseract==0.3.10
pytweening==1.0.7
pywin32-ctypes==0.2.0
requests==2.31.0
screeninfo==0.8.1
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
